+++ test/apps/_fastapi/__init__.py
@@ -1,0 +1,11 @@
+from fastapi import FastAPI
+
+
+def create_app():
+    app = FastAPI()
+
+    @app.get("/users")
+    async def root():
+        return {"success": True}
+
+    return app